backend/requirements.txt
Flask==2.0.3
flask_cors==3.0.10
Flask-Migrate==3.1.0
psycopg2==2.9.3
python-dotenv==0.19.2
spacy==3.2.4
Flask-RESTful==0.3.8
PyJWT==1.7.1
